Finish refactoring
