+++ src/Features/CSharp/Portable/ConvertProgram/ConvertToProgramMainCodeFixProvider.cs
@@ -17,6 +17,8 @@ using Microsoft.CodeAnalysis.Shared.Extensions;
 
 namespace Microsoft.CodeAnalysis.CSharp.ConvertProgram
 {
+    using static ConvertProgramTransform;
+
     [ExportCodeFixProvider(LanguageNames.CSharp, Name = PredefinedCodeFixProviderNames.ConvertToProgramMain), Shared]
     internal class ConvertToProgramMainCodeFixProvider : SyntaxEditorBasedCodeFixProvider
     {
@@ -52,7 +54,7 @@ namespace Microsoft.CodeAnalysis.CSharp.ConvertProgram
         protected override async Task FixAllAsync(
             Document document, ImmutableArray<Diagnostic> diagnostics, SyntaxEditor editor, CancellationToken cancellationToken)
         {
-            var fixedDocument = await ConvertProgramHelpers.ConvertToProgramMainAsync(document, cancellationToken).ConfigureAwait(false);
+            var fixedDocument = await ConvertToProgramMainAsync(document, cancellationToken).ConfigureAwait(false);
             var fixedRoot = await fixedDocument.GetRequiredSyntaxRootAsync(cancellationToken).ConfigureAwait(false);
 
             editor.ReplaceNode(editor.OriginalRoot, fixedRoot);
